Speed up execution by re-using unk info (#117)
Based on my benchmark, time was 1m23s before this change and 1m8s after.

The current code creates a new string and WordInfo object for every
single LatticeNode. The WordInfo is only used if the word is an unk, and
it's never modified. This change makes just one instance which all
LatticeNodes return if they need to.