Basic Cythonization
Unlike the other branch the tests pass on this one. Benchmark time went
down by a third compared to the previous commit.

I'm not sure the _c functions are necessary here - I think that's what
cpdef functions are for, but I had difficulty getting them working. Will
need to give that another look.