Fix connection cost lookup, fix #129 (#131)
* Fix connection cost lookup, fix #129

The connection costs lookup was backwards.

There was a comment in the pre-cython code that the call to the cost
lookup function looked backwards, but was actually correct. It was
calling with (l_node.right_id, r_node.left_id). I kept this order when I
replaced the function call with a memoryview access, but that was wrong;
I hadn't noticed that the access function was actually reversing the
order of its arguments.

This fix was verified by checking the tokenization of a short document
vs v0.4.5 and making sure there were no changes.

* Remove old comment